requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-LDAP3-Login==0.9.16
PyYAML==6.0.1
gitpython==3.1.42
python-ldap==3.4.4
Jinja2==3.1.3
ansible==9.3.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 